% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiv.BL16.fullRandom.modelRequire.R
\name{adaptiv.BL16.fullRandom.modelRequire}
\alias{adaptiv.BL16.fullRandom.modelRequire}
\title{adaptiv.BL16.fullRandom.modelRequire}
\usage{
adaptiv.BL16.fullRandom.modelRequire()
}
\value{
Kein expliziter Return-value
}
\description{
das yhat-modelRequire() fuer einen randomisierten BL32-Test
}
\details{
Es handelt sich hier um einen reinen Item-Sampler über den BL32-Item-Raum, der nach 16 Bällen abbricht.
}
\author{
Harald Fiedler
}

