% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiv.BL16.modelPredict.R
\name{adaptiv.BL16.modelPredict}
\alias{adaptiv.BL16.modelPredict}
\title{adaptiv.BL16.modelPredict}
\usage{
adaptiv.BL16.modelPredict(AnfrageDF)
}
\arguments{
\item{DF}{data.frame mit den Spalten:
\itemize{
  \item TestID: den Namen des Testformats. Beispielsweise: 'BL32'
  \item idS: die aktuelle Session-ID, beispielsweise: '002b6573-cf12-436d-bccd-0856b0bb0a25'
  \item idP: die Player-ID des jeweiligen Probanden, beispielsweise: 'fe553db4-bbde-43dd-a6a0-804b9e46c57'
  \item NamePlayer: Klarname des Spielers, beispielsweise: 'Mustermann, Tim'
  \item Birthday: Geburtsdatum des Spielers, beispielsweise: '2002-07-17'
  \item Team: Bezeichnung des Teams, z.B. "U17"
  \item SessionStart: der Zeitstempel für den Sessionsstart als String im Format "JJJJ-MM-TT HH:MM:SS", beispielsweise: '2014-03-22 13:42:03'
  \item adrB: eine Liste mit den Adressen der Ballkanonen, in der Reihenfolge ihrer Aktivierung, beispielsweise: [10, 45, 28] oder bei Sessionstart eine leere Liste '[]'.
  \item adrW: eine Liste von Listen mit den aufleuchtenden Zielfeldern. Beispielsweise: '[[4, 6], [21, 22], [50, 18]]' oder bei Sessionstart '[[]]'
  \item adrCol: eine Liste von Listen mit der Farbe der Zielfelder aus [[adrW]]. Die Farben werden alphabetisch in Großbuchstaben durchnummeriert. Beispielsweise: "[['A', 'B'], ['A', 'B'], ['B', 'A']]" oder bei Sessionstart '[[]]'
  \item adrOut: eine Liste mit den Adressen, wo der Ball tatsächlich raus ist. Beispielsweise: '[4, 23, 40]' oder bei Sessionstart '[]'
  \item FBt: eine Liste mit den Angaben über die Zeitdauer zwischen Ballkanoneneinwurf und Lichtschrankensignal beim Rausschießen des Balls (in Millisekunden), beispielsweise: '[2140, 2600, 8600]' oder bei Sessionstart '[]'
}}
}
\value{
Ein JSON-gültiger String mit den folgenden Informationen:
\itemize{
    \item TicketID ein stochastischer Identifikator für die REST-Request, z.B. "JTMfOgfTEiq6ZMxb"
    \item GameOver ein boolescher Wert der angibt, ob das Abbruchkriterium für die Testung erreicht wurde
    \item NextB Eine liste mit Informationen über den nächsten Ball
    \item Testergebnis Das Testergebnis, dass der Spieler auf der latenten Kompetenzdimension erhält. (Noch nicht implementiert, daher vorläufig NA)
    \item Seashell_output, Grafik im JSON format, wenn Gameover
}
}
\description{
modelPredict() fuer yhat-Modell eines adaptiven Samplers nach 5 Bällen des BL32-Testraums mit Stop nach 16 Bällen. Basiert auf dem Skript BL32MultiTargetSimTest.r (Teilcode wurde kopiert)
Der Output Ergebnis wurde nach dem adaptiv.BL16.fullRandom.modelPredict.R implementiert
}
\details{
Es handelt sich hier um die modelPredict-Funktion für yhat-Architektur. Bei dem Dienst handelt es sich um einen reinen Item-Sampler. Solange die Session-History weniger als 16 Bälle aufweist, wird aus dem Item-Raum der BL32 ein Item gewählt.
Der erste Ball ist immer BL01, die nächsten 4 Bälle werden ohne Zurücklegen gezogen.
}
\author{
Harald Fiedler, Kateryna Budzyak
}

