% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askDB.R
\name{askDB}
\alias{askDB}
\title{askDB}
\usage{
askDB(Anfangsdatum, Enddatum)
}
\arguments{
\item{Anfangsdatum}{character der Länge 1 im Format "JJJJ-MM-DD", welches dann zu einem Datumsobjekt umgewandelt wird. Achtung: Zeitzone könnte ein paar Probleme aufwerfen.}

\item{Enddatum}{character der Länge 1 im Format "JJJJ-MM-DD"}
}
\value{
data.frame das dann von SQL2DF() weiterverarbeitet werden kann.
}
\description{
Fragt die DB ab
}
\details{
Diese Funktion fragt auf localhost einen DB-dump der fbn-Datenbank ab und ersetzt das Copy&Paste-Verfahren der shinyApp
}
\author{
Harald Fiedler
}

