% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentRAW.R
\name{augmentRAW}
\alias{augmentRAW}
\title{augmentRAW}
\usage{
augmentRAW(SQL)
}
\arguments{
\item{SQL}{data.frame, dass per askDB() oder readRAW() eingelesen wurde}
}
\value{
data.frame
}
\description{
data.frame SQL wird angereichert
}
\details{
Es werden Derivate gebildet und Punktzahlen eingebunden
}
\author{
Harald Fiedler
}

