% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFiedler2012.R
\name{calcFiedler2012}
\alias{calcFiedler2012}
\title{calcFiedler2012}
\usage{
calcFiedler2012(SQL)
}
\arguments{
\item{SQL}{data.frame wie er per askDB() erzeugt wird.}
}
\value{
data.frame mit zusätzliche Spalte Fiedler2012
}
\description{
Berechnet einen Score nach der Heuristik von Fiedler (2012)
}
\details{
Es handelt sich dabei um die ursprüngliche Formel, die Harald Fiedler ehrenamtlich für CGoal am 12.2.2012 erstellt hat. Da CGoal inzwischen diese Formel mehrfach manipuliert hat, ist es wichtig, die Originalformel von CGoalScore zu unterscheiden.
}
\author{
Harald Fiedler
}

