% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFiedler2016a.R
\name{calcFiedler2016a}
\alias{calcFiedler2016a}
\title{calcFiedler2016a}
\usage{
calcFiedler2016a(SessionDF, ItemBank = readItemBank())
}
\arguments{
\item{SessionDF}{data.frame mit einer Session}

\item{ItemBank}{die ItemBank, wird per default mittels readItemBank() eingelesen.}
}
\value{
data.frame mit der Spalte für die IRT-Points
}
\description{
Berechnet für eine einzelne Session (DF) theta-Hat nach jedem Ball
}
\details{
DF darf nur eine Session beinhalten. Dann berechnet die Funktion thetaHat nach jedem Ball und stellt dies am Ende in der Spalte Fiedler2016a zur Verfügung.
}
\author{
Harald Fiedler
}

