% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectItemID.R
\name{detectItemID}
\alias{detectItemID}
\title{detectItemID}
\usage{
detectItemID(isMultiTarg, adrW, RW, AW, HW, sL, sR)
}
\arguments{
\item{adrW}{character, etwa "{13, 17}"}

\item{RW}{numeric der Länge 1, gibt den Reaktionswinkel an}

\item{AW}{numeric der Länge 1, gibt den Aktionswinkel an}

\item{HW}{character der Länge 1, etwa "FF"}

\item{sL}{numeric der Länge 1, eine natürliche Zahl zwischen 1 und 100}

\item{sR}{numeric der Länge 1, eine natürliche Zahl zwischen 1 und 100}

\item{isMulitTarg}{boolean der Länge 1, gibt an, ob das Item multiple Targets kennt}
}
\value{
charactger der Länge 1, z.B. c("BL03")
}
\description{
Liefert die ItemID eines Balls zurück
}
\details{
Liefert die ItemID eines Balls zurück, z.B. "BL03". Die Funktion ist nicht vektorwertig implementiert, sondern kann immer nur eine Abfrage auf einmal durchführen
}
\author{
Harald Fiedler
}

