% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectItemResponse.R
\name{detectItemResponse}
\alias{detectItemResponse}
\title{detectItemResponse}
\usage{
detectItemResponse(Stimulus)
}
\arguments{
\item{Stimulus}{ein data.frame mit den Spalten isMulitTarg, MultiTargs, RW, AW, HW, vA, sL und sR und einer Zeile. Es handelt sich also um eine Zeile aus DF, die einen Ball darstellt}
}
\value{
numeric der Länge 1
}
\description{
Liefert das Ergebnis eines Balls/Stimulus zurück, z.B. 0 oder 1
}
\details{
Liefert das Ergebnis eines Balls/Stimulus zurück, z.B. 0 oder 1. Aus dem data.frame ist nicht ersichtlich, welches für welches Modell die ItemResponse erhoben wird. Im dichotomen Rasch Modell wird das Ergebnis auf 0-1 codiert, während es für andere Modelle andere Erfassungen geben mag. Hier muss extern geklärt werden, welche ItemID welchem Modell zugeordnet ist.
}
\author{
Harald Fiedler
}

