% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/erstelleRaschMatrixSkeleton.R
\name{erstelleRaschMatrixSkeleton}
\alias{erstelleRaschMatrixSkeleton}
\title{erstelleRaschMatrixSkeleton}
\usage{
erstelleRaschMatrixSkeleton(DF, ItemIDNamen)
}
\arguments{
\item{DF}{data.frame auf Ballebene. Eine Spalte muss "idS".}

\item{character}{array mit den Itembezeichnungen, für die eine Rasch-Matrix erstellt werden soll.}
}
\value{
Eine Matrix voller NA, mit colnames=Itembezeichnungen und rownames=unique(idS)
}
\description{
erstellte eine NA-Matrix mit den Sessions als Zeilenindex und den Item-Namen als Spaltenindex
}
\details{
Achtung: eine Spalte des data.frame muss den Spaltennamen "idS" haben.
}
\examples{
rm(list=ls())
DF <- data.frame(c("SessionA", "SessionB"), c(22, 90), c(23, 18), c(10,12))
colnames(DF) <- c("idS", "It1", "It2", "It_von_wo_ganz_anders")
ItemIDNamen <- c("Item1", "Item2", "Item3")
print(DF)
erstelleRaschMatrixSkeleton(DF=DF, ItemIDNamen=ItemIDNamen)
}
\author{
Harald Fiedler
}

