% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fillRaschMatrixSkeleton.R
\name{fillRaschMatrixSkeleton}
\alias{fillRaschMatrixSkeleton}
\title{fillRaschMatrixSkeleton}
\usage{
fillRaschMatrixSkeleton(DF, RaschMatrixSkeleton)
}
\arguments{
\item{DF}{data.frame auf Ballebene, etwa per SQL2DF erworben}

\item{RaschMatrixSkeleton}{matrix , belabeled mit SessionIDs und ItemIDs, wird etwa aus erstelleRaschMatrixSkeleton() erworben.}
}
\value{
Eine RaschMatrix mit vielen NA, und wenigen 0 und einigen 1en.
}
\description{
Füllt die NA-Matrix mit 0 und 1, wo es zutreffend ist.
}
\author{
Harald Fiedler
}

