% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAdrWAlsListe.R
\name{getAdrWAlsListe}
\alias{getAdrWAlsListe}
\title{getAdrWAlsListe}
\usage{
getAdrWAlsListe(adrW)
}
\arguments{
\item{adrW}{character}
}
\value{
list mit numerischen Elementen
}
\description{
adrW-Einträge als numerische Liste
}
\details{
Aus c("{1, 2, 3, 4}", "{11, 12, 13, 14}") mach mache list(c(1, 2, 3, 4), c(11, 12, 13, 14))
}
\author{
Harald Fiedler
}

