% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFirstAdrW.R
\name{getFirstAdrW}
\alias{getFirstAdrW}
\title{getFirstArdW}
\usage{
getFirstAdrW(adrW)
}
\arguments{
\item{adrW}{character Vektor, etwa c("{2, 4, 5}", "{12,19}", "{4}")}
}
\value{
numeric
}
\description{
Gibt erstes Ziel in adrW
}
\details{
Unabhängig davon, ob in adrW ein multiTarget oder singleTarget-Design hinterlegt wird, wird das erste Ziel zurückgegeben.
}
\author{
Harald Fiedler
}

