% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHW.R
\name{getHW}
\alias{getHW}
\title{getHW}
\usage{
getHW(SQL)
}
\arguments{
\item{SQL}{data.frame welches durch read.csv() einer SQL-Query entnommen wurde}
}
\value{
character mit Einträgen aus c("FF", "FH", "HF", "HH"), wobei FF=Flach Flach bedeutet und HH=Hoch Hoch.
}
\description{
Hilfsfunktion von SQL2DF zur Ermittlung von Höhenwinkel FF-FH-HF-HH
}
\details{
Je nachdem ob eine obere Ballkanonen oder eine untere Ballkanone zum Zuge kommt, oder ein unteres Zeil respektive oberes Ziel, kommt ein anderer Höhenwinkle zu stande.
}
\author{
Harald Fiedler
}

