% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getItemICC.R
\name{getItemICC}
\alias{getItemICC}
\title{getItemICC}
\usage{
getItemICC(DF, ItemBank, MaximaleToleranz = 10)
}
\arguments{
\item{DF}{data.frame Mittels SQL2DF(SQL=SQL) erzeugt wird}

\item{ItemBank}{data.frame Wird etwa durch ItemBank=readItemBank() gewonnen}

\item{MaximaleToleranz}{numeric der Länge 1, gibt an, wie viel sL.x von sL.y bzw. sR.x von sR.y abweichen darf, damit das Item in DF identifiziert wird mit dem Item aus der Itembank. Dabei stammt *.x aus DF und *.y aus der ItemBank. Default ist 10.}
}
\value{
DF wird um die Spalten ItemID, ICCa, ICCb, ICCc und ICCd angereichert.
}
\description{
Fügt die Spalten ItemID, ICCa, ICCb, ICCc, ICCd
}
\details{
Aus einem DF (erzeugt mittels SQL2DF(SQL = SQL)) und der ItemBank wird ein Merge erzeugt. Dabei nutze ich AW, RW, HW und vA als Key. In einem zweiten Schritt wird die Identifikation der Items gelöscht, wenn zwischen dem tstsächlichen sL bzw. sR und dem in der ItemBank hinterlegten sL und sR eine zu große Diskrepanz entsteht.
}
\examples{
Pfad <- system.file("extdata", package="Rbonaut2", "Footbonaut_Datenabfrage_RicoWehrle.csv")
SQL <- read.csv(file=Pfad, sep=";", header=TRUE, encoding="utf8", stringsAsFactors = FALSE)
DF <- SQL2DF(SQL = SQL)
ItemBank=readItemBank()
F14 <- getItemICC(DF=DF, ItemBank=readItemBank(), MaximaleToleranz=15)
head(F14)
}
\author{
Harald Fiedler
}

