% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNachname.R
\name{getNachname}
\alias{getNachname}
\title{getNachname}
\usage{
getNachname(Spielername)
}
\arguments{
\item{Spielername}{character Vektor von beliebiger Länge}
}
\value{
character Vektor der gleichen Länge wie der an die Funktion übergebene Vektor
}
\description{
Gibt aus einem Spielername den Vornamen
}
\details{
Spielernamen können in SQL-Abfragen des FBN beispielsweise "Dogan, Isa" sein. Es wird "Isa" zurückgegeben.
}
\author{
Harald Fiedler
}

