% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/implodeFilledRaschMatrixSkeleton4Quality.R
\name{implodeRaschMatrix4Quality}
\alias{implodeRaschMatrix4Quality}
\title{implodeFilledRaschMatrixSkeleton4Quality}
\usage{
implodeRaschMatrix4Quality(RaschMatrixSkeletonFilled, MissingToleranz = 0.1)
}
\arguments{
\item{RaschMatrixSkeletonFilled}{matrix bestehend aus vielen NA und einigen 0 und 1en, wie man sie aus fillRaschMatrixSkeleton() erwirbt}

\item{MissingToleranz}{numeric der Länge 1. Wie viel Prozent fehlende Bälle werden höchstens erlaubt. Default ist 10 Prozent}
}
\value{
matrix bestehend aus 0en und 1en und ganz ganz wenigen NA. Die Spalten tragen Item-Namen, und die Zeilen die Probandennamen, hier: idS
}
\description{
Lässt alle Probanden/Sessions weg, die zu viele NA haben, um eine sinnvolle ItemAnylse durchzuführen.
}
\author{
Harald Fiedler
}

