% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isMultiTarget.R
\name{isMultiTarget}
\alias{isMultiTarget}
\title{isMultiTarget}
\usage{
isMultiTarget(adrW)
}
\arguments{
\item{adrW}{character Array, etwa c("{22, 33, 44, 55}", "{11, 22222, 11111}", "{99}")}
}
\value{
boolescher Vektor
}
\description{
Hilfsfunktion von SQL2DF(): ist adrW multitargetting?
}
\details{
Sagt, ob {11, 21, 16} oder {23} unter adrW abgespeichert wurde
}
\author{
Harald Fiedler
}

