% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/istFormatNachnameKommaVorname.R
\name{istFormatNachnameKommaVorname}
\alias{istFormatNachnameKommaVorname}
\title{istFormatNachnameKommaVorname}
\usage{
istFormatNachnameKommaVorname(Spielername)
}
\arguments{
\item{Spielername}{String}
}
\value{
Boolescher Wert, der angibt, ob das Format Name, Vorname (mutmaßlich) vorliegt
}
\description{
Ist ein Eintrag vom Format "Fiedler, Harald"
}
\details{
In den SQL-Auszügen des FBN finden sich Spielername vom Format "Fiedler, Harald", aber auch "A_TEST_Forschung". Die Funktion testet komponentenweise, ob zwei Strings kommagetrennt gepastet sind.
}
\author{
Harald Fiedler
}

