% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isUTF8.R
\name{isUTF8}
\alias{isUTF8}
\title{isUTF8}
\usage{
isUTF8(file = "*")
}
\arguments{
\item{file}{character der L채nge 1 der angibt, ob sch채tzungsweise UTF8 vorliegt}
}
\value{
siehe *details*
}
\description{
sch채tzt, ob UTF8 Codierung vorliegt
}
\details{
Diese Funktion funktioniert nur auf Mac, wobei hierzu per *brew install moreutils* installiert sein muss. Sie gibt character(0) zur체ck, wenn kein utf8-nonkonformes Zeichen gefunden wurde, und ansonsten eine Liste mit Angaben zu invaliden Zeichen
}
\examples{
isUTF8(file="~/Desktop")
}
\author{
Harald Fiedler
}

