% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormTree6.16.R
\docType{data}
\name{NormTree6.16}
\alias{NormTree6.16}
\title{NormTree6.16 ist eine verschachtelte Liste}
\format{A nested list about 1194 players}
\usage{
data(NormTree6.16)
}
\description{
NormTree6.16
|___ Altersgruppen/Mannschaften
     |__ RAW : data.frame mit SessionID, Fiedler2016a, FBt, FBq
     |__ HIST : list of counts and breaks
     |__ Kernel
         |__FBq
         |__FBt
         |__Level
         |__FBqFBt
}
\keyword{datasets}

