% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotFBN.R
\name{plotFBN}
\alias{plotFBN}
\title{plotFBN}
\usage{
plotFBN(Adresses = TRUE)
}
\arguments{
\item{Adresses}{boolescher Wert der angibt, ob die Fensteradressen mit eingegeben werden sollen.}
}
\description{
plotFBN
}
\details{
Zeichnet schematisch den Footbonaut
}
\examples{
plotFBN(Adresses=FALSE)
}
\author{
Harald Fiedler
}

