% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeaShell.R
\name{plotSeaShell}
\alias{plotSeaShell}
\title{plotSeaShell}
\usage{
plotSeaShell(x, A, B, TitelA, TitelB, developperMode = FALSE,
  gameover = FALSE)
}
\arguments{
\item{x}{numeric Der Item-Response-Wert des Probanden}

\item{A}{numeric Die Item-Response-Werte von Gruppe A}

\item{B}{numeric die Item-Response-Werte von Gruppe B}

\item{TitelA}{character mit der Headline für Gruppe A}

\item{TitelB}{character mit der Headline für Gruppe B}

\item{developperMode}{boolescher Wert, der eine Augenscheinkontrolle des Kernels und der Histogramms erlaubt.}
}
\value{
Die Funktion hat keinen Ausgabe-wert
}
\description{
Plottet die SeaShellGrafik
}
\details{
Es handelt sich um die extra für die Normgruppenvergleiche von Hoffenheim entwickelte Grafik
}
\author{
Harald Fiedler
}

