% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAUGMENTED.R
\name{readAUGMENTED}
\alias{readAUGMENTED}
\title{readAUGMENTED}
\usage{
readAUGMENTED(Dateiname)
}
\arguments{
\item{Dateiname}{character der Länge 1 mit Dateiname (ohne Endung). Der Pfad wird automatisch auf die Dropbox gesetzt, genauer in den Ordner RAW vom Ordner Hoffenheim}
}
\description{
Liest die per writeAUGMENTED() gespeicherten Dateien ein
}
\examples{
#Dateiname = "RAW-2015-04"
#head(readAUGMENTED(Dateiname = Dateiname))

}
\author{
Harald Fiedler
}

