% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readItemBank.R
\name{readItemBank}
\alias{readItemBank}
\title{readItemBank}
\usage{
readItemBank(file = "default")
}
\arguments{
\item{Pfad}{character der L채nge 1, der den Pfadname zu einer .csv-Datei darstellt. Die Datei muss eine g체ltige ItemBank im Sinne des 4-PL-Modells sein.
Als Default-Wert f체r den Pfad fungiert ein Pfad zu einer Pakte-Datei, die in der Lib installiert wurde (was der eigentliche Clou dieser Funktion ist).}
}
\value{
data.frame f체r das 4PL-Modell
}
\description{
Liest die ItemBank ein
}
\details{
Es wird die ItemBank im 1:4PL-Modell eingelesen.
}
\author{
Harald Fiedler
}

