% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRAW.R
\name{readRAW}
\alias{readRAW}
\title{readRAW}
\usage{
readRAW(Dateiname)
}
\arguments{
\item{Dateiname}{character der Länge 1 mit Dateiname (ohne Endung). Der Pfad wird automatisch auf die Dropbox gesetzt, genauer in den Ordner RAW vom Ordner Hoffenheim}
}
\value{
data.frame der einer SQL-Abfrage gleicht.
}
\description{
Liest die per writeRAW() gespeicherten Dateien ein
}
\examples{
Dateiname = "RAW-2015-04"
head(readRAW(Dateiname = Dateiname))

}
\author{
Harald Fiedler
}

