% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQL2DF.R
\name{SQL2DF}
\alias{SQL2DF}
\title{SQL2DF}
\usage{
SQL2DF(SQL)
}
\arguments{
\item{SQL}{data.frame, dass per read.csv eingelesen wurde}
}
\value{
data.frame
}
\description{
SQL zu data.frame
}
\details{
Mit shinySQL erhalten wir von CGoal SQL-Abfragen händisch als .csv-Files zurück. Diese werden in ein data.frame umgewandelt
}
\author{
Harald Fiedler
}

