% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRAW.R
\name{writeRAW}
\alias{writeRAW}
\title{writeRAW}
\usage{
writeRAW(DF, Dateiname)
}
\arguments{
\item{DF}{data.frame der per askDB() erzeugte Datensatz}

\item{Dateiname}{character der Länge 1, gibt den Dateinamen ohne Endung an. Der Pfad ist hard coded zur Dropbox}
}
\description{
Der per askDB() erzeugte Datensatz (ein data.frame) wird als .csv in der Dropbox abgespeichert.
}
\author{
Harald Fiedler
}

