% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectItemID.R
\name{detectItemID}
\alias{detectItemID}
\title{detectItemID}
\usage{
detectItemID(Stimulus)
}
\arguments{
\item{Stimulus}{ein data.frame mit den Spalten isMulitTarg, MultiTargs, RW, AW, HW, vA, sL und sR und einer Zeile. Es handelt sich also um eine Zeile aus DF, die einen Ball darstellt}
}
\value{
charactger der L채nge 1, z.B. c("BL03")
}
\description{
Liefert die ItemID eines Balls/Stimulus zur체ck
}
\details{
Liefert die ItemID eines Balls/Stimulus zur체ck, z.B. "BL03". Die Funktion ist nicht vektorwertig implementiert, sondern kann immer nur eine Abfrage auf einmal durchf체hren
}
\author{
Harald Fiedler
}

