% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNachname.R
\name{getNachname}
\alias{getNachname}
\title{getNachname}
\usage{
getNachname(Spielername)
}
\arguments{
\item{Spielername}{character Vektor von beliebiger Länge}
}
\value{
character Vektor der gleichen Länge wie der an die Funktion übergebene Vektor
}
\description{
Hilfsfunktion von augmentRWA(): Gibt aus einem Spielername den Vornamen
}
\details{
Spielernamen können in SQL-Abfragen des FBN beispielsweise "Dogan, Isa" sein. Es wird "Isa" zurückgegeben.
}
\examples{
Spielername <- c("Fiedler, Harald", "Mayer, Jan", "A-Team")
getNachname(Spielername = Spielername)
}
\author{
Harald Fiedler
}

