% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSessionTimeStamp.R
\name{getSessionTimeStamp}
\alias{getSessionTimeStamp}
\title{getSessionTimeStamp}
\usage{
getSessionTimeStamp(DatumString)
}
\arguments{
\item{DatumString}{String, etwa "2015-08-27 18:59:25.328383+02"}
}
\value{
Ein String, etwas "18:59:25"
}
\description{
Hilfsfunktion von SQL2DF(): ermittelt Sessionstart
}
\details{
Macht aus 2015-08-27 18:59:25.328383+02 den String 18:59:25
}
\examples{
DatumString <- c("2015-08-27 18:59:25.328383+02", "2015-08-27 18:59:25.328383+02", "2015-08-27 18:59:25.328383+02")
}
\author{
Harald Fiedler
}

