% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isUTF8.R
\name{isUTF8}
\alias{isUTF8}
\title{isUTF8}
\usage{
isUTF8(file, echo = TRUE)
}
\arguments{
\item{file}{character der Länge 1 der angibt, ob schätzungsweise UTF8 vorliegt}

\item{echo}{boolean der Länge 1 sagt, ob der Systemoutput angezeigt werden soll}
}
\value{
siehe *details*
}
\description{
schätzt, ob UTF8 Codierung vorliegt
}
\details{
Diese Funktion funktioniert nur auf Mac, wobei hierzu per *brew install moreutils* installiert sein muss. Sie gibt character(0) zurück, wenn kein utf8-nonkonformes Zeichen gefunden wurde, und ansonsten eine Liste mit Angaben zu invaliden Zeichen
}
\examples{
isUTF8(file="~/Desktop")
}
\author{
Harald Fiedler
}

