% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/playedAngle.R
\name{playedAngle}
\alias{playedAngle}
\title{playedAngle}
\usage{
playedAngle(adrA, adrB, Clockwise = FALSE)
}
\arguments{
\item{Clockwise}{boolean der angibt, ob die Winkel im Uhrzeigersinn oder im mathematisch positiven Sinn gemessen werden. Default ist der mathematisch positive Sinn.}

\item{A}{numeric Adressen der Ausgangsfelder}

\item{B}{numeric Adressen der Zielfelder}
}
\value{
numeric mit Winkel aus 0:180
}
\description{
Winkel zwischen zwei Adressen
}
\details{
Gibt den Winkel zwischen zwei FBN-Adressen
}
\author{
Harald Fiedler
}

